Ledger.Commodity

Documentation

type Commodity = Key

Commodities are simply indices into a commodity info map, which relates such commodities to the information known about them.

data CommodityInfo

All of the information known about a commodity.

Constructors

CommodityInfo

Fields

commSymbol :: !Text

commPrecision :: !Int

commSuffixed :: !Bool

commSeparated :: !Bool

commThousands :: !Bool

commDecimalComma :: !Bool

commNoMarket :: !Bool

commBuiltin :: !Bool

commKnown :: !Bool

commPrimary :: !Bool

commConversions :: !(IntMap (Map UTCTime Rational))

defaultCommodityInfo :: CommodityInfo

Return a CommodityInfo with defaults selected for all fields. It is intended that at least one field of the result will be modified immediately.

data CommodityMap

A commodities map, relating commodity indices to information about those commodities.

Constructors

CommodityMap

Fields

commodities :: IntMap CommodityInfo

findConversion

Arguments

:: Commodity
Source commodity

-> Commodity
Target commodity

-> UTCTime
Look for conversions on or before this

-> CommodityMap
Set of commodities to search

-> Maybe (UTCTime, Rational)

Lookup a price conversion from the source commodity to the target, using data from the given time or earlier. Result is Nothing if no conversion can be found, or else the best conversion ratio plus the time of the oldest link.

addConversion :: Commodity -> Commodity -> UTCTime -> Rational -> State CommodityMap ()

Add a price conversion in the form of a ratio between two commodities at a specific point in time.

intAStar

Arguments

:: (Ord c, Num c)

=> (Key -> IntMap c)
The graph we are searching through, given as a function from vertices to their neighbours.

-> (Key -> c)
Heuristic distance to the (nearest) goal. This should never overestimate the distance, or else the path found may not be minimal.

-> (Key -> Bool)
The goal, specified as a boolean predicate on vertices.

-> Key
The vertex to start searching from.

-> Maybe [Key]
An optimal path, if any path exists. This excludes the starting vertex.

This function computes an optimal (minimal distance) path through a graph in a best-first fashion, starting from a given starting point.

intAStarM

Arguments

:: (Monad m, Ord c, Num c)

=> (Key -> m (IntMap c))
The graph we are searching through, given as a function from vertices to their neighbours.

-> (Key -> m c)
Heuristic distance to the (nearest) goal. This should never overestimate the distance, or else the path found may not be minimal.

-> (Key -> m Bool)
The goal, specified as a boolean predicate on vertices.

-> m Key
The vertex to start searching from.

-> m (Maybe [Key])
An optimal path, if any path exists. This excludes the starting vertex.

This function computes an optimal (minimal distance) path through a graph in a best-first fashion, starting from a given starting point.